requirements.txt
aerich==0.6.3
aiogram==2.20
aiohttp==3.8.1
aioredis==2.0.1
aioschedule==0.5.2
aiosignal==1.2.0
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
Babel==2.9.1
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
dictdiffer==0.9.0
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
iso8601==1.0.2
isort==5.10.1
Mako==1.2.0
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pydantic==1.9.1
pypika-tortoise==0.1.5
python-dotenv==0.20.0
pytz==2022.1
tomli==2.0.1
tomlkit==0.11.0
tortoise-orm==0.19.1
typing_extensions==4.2.0
yarl==1.7.2
zipp==3.8.0
